aider: Updated 'aider.add' command to get the currently selected file in VS Code and send the "/add <filename>" command to the Aider process.
# Aider chat conversation:

USER: When the "Add File to Aider" command is run, send the command "/add <filename>" to aider, where <filename> is the filename of the currently selected file in vscode.

ASSISTANT: To make this change we need to modify `vscode-aider-extension/extension.js`:

1. Update the 'aider.add' command to get the currently selected file in VS Code and send the "/add <filename>" command to the Aider process.

Here is the *SEARCH/REPLACE* block:

vscode-aider-extension/extension.js
```javascript
<<<<<<< SEARCH
    disposable = vscode.commands.registerCommand('aider.add', function () {
        // The code you place here will be executed every time your command is executed
        // TODO: Implement add functionality
    });
=======
    disposable = vscode.commands.registerCommand('aider.add', function () {
        // The code you place here will be executed every time your command is executed
        // Get the currently selected file in VS Code
        let activeEditor = vscode.window.activeTextEditor;
        if (!activeEditor) {
            return; // No open text editor
        }
        let filePath = activeEditor.document.fileName;

        // Send the "/add <filename>" command to the Aider process
        if (aiderProcess) {
            aiderProcess.stdin.write(`/add ${filePath}\n`);
        }
    });
>>>>>>> REPLACE
```

Please note that this implementation assumes that the Aider process is reading commands from its standard input. If the Aider process is expecting commands to be sent in a different way, this code will need to be adjusted accordingly.

USER: InvalidEditBlock: edit failed!

vscode-aider-extension/extension.js does not contain the *exact chunk* of SEARCH lines you specified.
Try again.
DO NOT skip blank lines, comments, docstrings, etc!
The SEARCH block needs to be EXACTLY the same as the lines in vscode-aider-extension/extension.js with nothing missing!

vscode-aider-extension/extension.js does not contain these 4 exact lines in a row:
```
    disposable = vscode.commands.registerCommand('aider.add', function () {
        // The code you place here will be executed every time your command is executed
        // TODO: Implement add functionality
    });
```

ASSISTANT: I apologize for the confusion. Let's try again with the correct *SEARCH/REPLACE* block.

Here is the correct *SEARCH/REPLACE* block:

vscode-aider-extension/extension.js
```javascript
<<<<<<< SEARCH
    let disposable = vscode.commands.registerCommand('aider.add', function () {
        // The code you place here will be executed every time your command is executed
        // TODO: Implement add functionality
    });
=======
    let disposable = vscode.commands.registerCommand('aider.add', function () {
        // The code you place here will be executed every time your command is executed
        // Get the currently selected file in VS Code
        let activeEditor = vscode.window.activeTextEditor;
        if (!activeEditor) {
            return; // No open text editor
        }
        let filePath = activeEditor.document.fileName;

        // Send the "/add <filename>" command to the Aider process
        if (aiderProcess) {
            aiderProcess.stdin.write(`/add ${filePath}\n`);
        }
    });
>>>>>>> REPLACE
```

Please note that this implementation assumes that the Aider process is reading commands from its standard input. If the Aider process is expecting commands to be sent in a different way, this code will need to be adjusted accordingly.