aider: Added configuration option for regexes that match filenames which won't be automatically added to the Aider terminal process.
# Aider chat conversation:

ASSISTANT: To add a configuration option for regexes that match filenames which won't be automatically added to the Aider terminal process, we need to modify two files:

1. `vscode-aider-extension/package.json` to define the configuration option.
2. `vscode-aider-extension/extension.js` to use the configuration option when deciding whether to add a file to the Aider terminal process.

Here are the *SEARCH/REPLACE* blocks:

vscode-aider-extension/package.json
```json
<<<<<<< SEARCH
    "contributes": {
        "commands": [
=======
    "contributes": {
        "configuration": {
            "title": "Aider",
            "properties": {
                "aider.ignoreFiles": {
                    "type": "array",
                    "default": [],
                    "items": {
                        "type": "string"
                    },
                    "description": "List of regexes that match filenames that won't automatically be added to the Aider terminal process."
                }
            }
        },
        "commands": [
>>>>>>> REPLACE
```

vscode-aider-extension/extension.js
```javascript
<<<<<<< SEARCH
    vscode.window.onDidChangeActiveTextEditor((editor) => {
        if (editor) {
            let filePath = editor.document.fileName;
            if (terminal) {
                terminal.sendText(`/add ${filePath}`);
            }
        }
    });
=======
    vscode.window.onDidChangeActiveTextEditor((editor) => {
        if (editor) {
            let filePath = editor.document.fileName;
            let ignoreFiles = vscode.workspace.getConfiguration('aider').get('ignoreFiles');
            let shouldIgnore = ignoreFiles.some((regex) => new RegExp(regex).test(filePath));
            if (!shouldIgnore && terminal) {
                terminal.sendText(`/add ${filePath}`);
            }
        }
    });
>>>>>>> REPLACE
```

In the first *SEARCH/REPLACE* block, we add a new configuration option "aider.ignoreFiles" to the package.json file. This option is an array of strings, where each string is a regex that matches filenames to ignore.

In the second *SEARCH/REPLACE* block, we modify the `onDidChangeActiveTextEditor` event handler in the extension.js file. Before adding a file to the Aider terminal process, we check if the filename matches any of the regexes in the "aider.ignoreFiles" configuration option. If it does, we don't add the file.