aider: Allow JavaScript files to be compiled and include all files in the src directory and its subdirectories in tsconfig.json.
# Aider chat conversation:

USER: Now I'm getting the error: error TS18003: No inputs were found in config file '/Users/mflower/code/vscode-aider/vscode-aider-extension/tsconfig.json'. Specified 'include' paths were '["**/*"]' and 'exclude' paths were '["node_modules",".vscode-test"]'.

ASSISTANT: The error message is indicating that the TypeScript compiler is not finding any input files to compile based on the configuration in your `tsconfig.json` file.

The `include` and `exclude` properties in the `tsconfig.json` file specify which files the TypeScript compiler should include and exclude when it compiles your code. By default, if no `include` or `exclude` property is specified, the compiler includes all TypeScript (`.ts`, `.d.ts`, and `.tsx`) files in the project, excluding those in `node_modules`, `bower_components`, `jspm_packages`, `.tmp`, and the `outDir` directory.

In your case, the `include` property is set to `["**/*"]`, which means "include all files in all directories". The `exclude` property is set to `["node_modules", ".vscode-test"]`, which means "exclude all files in the `node_modules` and `.vscode-test` directories".

The error message is saying that, even with these settings, the compiler is not finding any files to compile. This could be because:

1. There are no TypeScript files in your project. If your project only contains JavaScript files, you need to either convert them to TypeScript or configure the TypeScript compiler to compile JavaScript files by setting the `allowJs` property to `true` in your `tsconfig.json` file.

2. The TypeScript files in your project are located in directories that are not being included by the `include` property. If this is the case, you need to update the `include` property to include these directories.

3. The TypeScript files in your project are being excluded by the `exclude` property. If this is the case, you need to update the `exclude` property to not exclude these files.

To fix the issue, you need to modify the `tsconfig.json` file. Here is the *SEARCH/REPLACE* block:

vscode-aider-extension/tsconfig.json
```json
<<<<<<< SEARCH
{
    "compilerOptions": {
        "module": "commonjs",
        "target": "es6",
        "outDir": "out",
        "lib": ["es6"],
        "sourceMap": true,
        "rootDir": "src",
        "strict": true,
        "esModuleInterop": true
    },
    "exclude": ["node_modules", ".vscode-test"]
}
=======
{
    "compilerOptions": {
        "module": "commonjs",
        "target": "es6",
        "outDir": "out",
        "lib": ["es6"],
        "sourceMap": true,
        "rootDir": "src",
        "strict": true,
        "esModuleInterop": true,
        "allowJs": true
    },
    "include": ["src/**/*"],
    "exclude": ["node_modules", ".vscode-test"]
}
>>>>>>> REPLACE
```

This change will allow the TypeScript compiler to compile JavaScript files and will include all files in the `src` directory and its subdirectories.